# doc/api.yaml
openapi: 3.0.0
paths:
...
  /album:
    post:
      description: This endpoint creates a new album with a name, year, and an
        optional artist ID.
      operationId: AlbumsController_addAlbum
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: The album has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad Request. The input data is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - year must not be greater than 2025
                    - year must not be less than 1900
                    - year must be an integer number
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Artist not exist. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: &a5
        - access-token: []
      summary: Create a new album
      tags: &a6
        - Albums
    get:
      description: This endpoint retrieves a list of all albums.
      operationId: AlbumsController_getAlbums
      parameters: []
      responses:
        "200":
          description: A list of albums has been successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
      security: *a5
      summary: Get all albums
      tags: *a6
  /album/{id}:
    get:
      description: This endpoint retrieves an album by its ID.
      operationId: AlbumsController_getAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "200":
          description: The album has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
      security: *a5
      summary: Get an album by ID
      tags: *a6
    put:
      description: This endpoint updates an existing album by its ID with new data.
      operationId: AlbumsController_updateAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: The The user password has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
        "422":
          description: Artist not exist. The artist with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Artist with ID 123e4567-e89b-12d3-a456-426614174001 does not exist
                  error: Not Found
      security: *a5
      summary: Update an album by ID
      tags: *a6
    delete:
      description: This endpoint deletes an album by its ID.
      operationId: AlbumsController_deleteAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "204":
          description: The album has been successfully deleted.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found
                  error: Not Found
      security: *a5
      summary: Delete an album by ID
      tags: *a6
...
  /favs/album/{id}:
    post:
      description: This endpoint adds a album to the favorites.
      operationId: FavoritesController_addAlbumToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "201":
          description: The album has been successfully added to favorites.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddedFavorite"
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "422":
          description: Album not exist. The album with the specified ID does not exist.
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 does not exist
                  error: Not Found
      security: *a9
      summary: Add a album to favorites
      tags: *a10
    delete:
      description: This endpoint removes a album from the favorites.
      operationId: FavoritesController_removeAlbumFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the album (UUID v4 format)
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174002
            type: string
      responses:
        "204":
          description: The album has been successfully removed from favorites.
        "400":
          description: Bad Request. The ID property is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message: Invalid ID format. Must be a UUID v4
                  error: Bad Request
        "401":
          description: Unauthorized. The user is not authenticated or the token is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized. The user is not authenticated or the token is invalid.
                  error: Unauthorized
        "404":
          description: Album not found in favorites. The album with the specified ID does
            not exist in favorites.
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Album with ID 123e4567-e89b-12d3-a456-426614174002 not found in
                    favorites
                  error: Not Found
      security: *a9
      summary: Remove a album from favorites
      tags: *a10
...
  /auth/signup:
    post:
      description: This endpoint allows a new user to sign up with a login and password.
      operationId: AuthController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: The user has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request. The input data is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - login must be longer than or equal to 3 characters
                    - password must be longer than or equal to 4 characters
                  error: Bad Request
      summary: Sign up a new user
      tags: &a11
        - Auth
  /auth/login:
    post:
      description: This endpoint allows an existing user to log in with their login
        and password.
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "200":
          description: Tokens has been successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad Request. The input data is invalid.
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - login must be longer than or equal to 3 characters
                    - password must be longer than or equal to 4 characters
                  error: Bad Request
        "403":
          description: Forbidden. The user does not have permission to perform this action.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Login or password is incorrect
                  error: Forbidden
      summary: Log in an existing user
      tags: *a11
  /auth/refresh:
    post:
      description: This endpoint allows a user to refresh their session using a
        refresh token.
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: Tokens has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Refresh token not provided
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Refresh token not provided
                  error: Unauthorized
        "403":
          description: Forbidden. The user does not have permission to perform this action.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Refresh token is invalid or expired
                  error: Forbidden
      summary: Refresh user session
      tags: *a11
info:
  title: Home Library Service
  description: API for managing music albums, artists, and tracks
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      description: Enter JWT access token
      in: header
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: The login of the user
          example: john_doe
          minLength: 3
          maxLength: 20
        password:
          type: string
          description: The password of the user
          example: password
          minLength: 4
          maxLength: 100
      required:
        - login
        - password
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        login:
          type: string
          example: john_doe
        version:
          type: number
          example: 1
        createdAt:
          format: int64
          type: integer
          example: 1716927665136
        updatedAt:
          format: int64
          type: integer
          example: 1716927665136
...
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the album
          example: Abbey Road
        year:
          type: number
          description: The year the album was released
          example: 1969
        artistId:
          type: string
          description: The artist ID associated with the album
          nullable: true
          format: uuid
      required:
        - name
        - year
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        name:
          type: string
          example: Abbey Road
        year:
          type: number
          description: The year the album was released
          example: 1969
        artistId:
          type: string
          description: The artist ID associated with the album
          nullable: true
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the album
          example: Abbey Road
        year:
          type: number
          description: The year the album was released
          example: 1969
        artistId:
          type: string
          description: The artist ID associated with the album
          nullable: true
          format: uuid
...
    AddedFavorite:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
          description: The unique identifier of the added favorite entity
        type:
          type: string
          enum:
            - artist
            - album
            - track
          example: artist
          description: The type of the added favorite entity
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          format: token
          description: JWT access token for authenticated requests
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          format: token
          description: JWT refresh token for obtaining new access tokens
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: JWT refresh token for obtaining new access tokens
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          format: token
      required:
        - refreshToken