# src/auth/interfaces/auth-response.interface.ts
import { ApiPropertyOptional } from '@nestjs/swagger';

export class AuthResponse {
  @ApiPropertyOptional({
    format: 'token',
    description: 'JWT access token for authenticated requests',
    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...',
  })
  accessToken: string;

  @ApiPropertyOptional({
    format: 'token',
    description: 'JWT refresh token for obtaining new access tokens',
    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...',
  })
  refreshToken: string;
}